Added Windows build support.
+ updated makefiles
+ some code changes
+ Need to fix compatability issues (all those warnings...)
+ add bat files

+++ src/dsl-manager.c
@@ -70,7 +70,7 @@ void dsl_lua_load_scripts() {
 
     ids = gff_get_id_list(DSLDATA_GFF_INDEX, GFF_MAS);
     amt = gff_get_resource_length(DSLDATA_GFF_INDEX, GFF_MAS);
-    printf("Detected %ld master GPL files.\n", amt);
+    printf("Detected %ld master GPL files.\n", (long int)amt);
     for (i = 0; i < amt; i++) { mas_max = mas_max > ids[i] ? mas_max : ids[i]; }
     mas_max++;
     mas_scripts = malloc(sizeof(char*) * mas_max);
@@ -79,7 +79,7 @@ void dsl_lua_load_scripts() {
 
     ids = gff_get_id_list(DSLDATA_GFF_INDEX, GFF_GPL);
     amt = gff_get_resource_length(DSLDATA_GFF_INDEX, GFF_GPL);
-    printf("Detected %ld standard GPL files.\n", amt);
+    printf("Detected %ld standard GPL files.\n", (long int)amt);
     for (i = 0; i < amt; i++) { gpl_max = gpl_max > ids[i] ? gpl_max : ids[i]; }
     gpl_max++;
     gpl_scripts = malloc(sizeof(char*) * gpl_max);
